+++ spec/factories/alpha.rb
@@ -6,7 +6,7 @@ FactoryBot.define do
 
     values = samples.each_with_object(Hash.new) do |value, hash|
       hex = "%02X" % value
-      range_01  = (value / 255.to_f).round(5)
+      range_01  = (value / 255.to_f).round(3)
       hash[hex] = {rgb: range_01, rgb_255: value}
     end