+++ NEWS.md
@@ -1,4 +1,9 @@
 # News
 
 ## 1.0.0 (September 13, 2021)
+
   * First version
+  
+## 1.1.0 (September 14, 2021)
+
+  * `::hex_to_rgb` Now supports short version of HEX and rounding for the alpha channel